move batchnorm and layernorm fusion to decomposition
